moved images to correct position and changed the import links to match aswell as removed  duplicate logo
