+++ client/public/trash-can.png
